Fix zero or negative auto barWidth and allow minimum value
This PR firstly fixes a problem whereby the auto-calculated barWidth can
be zero or negative, if there are a high number of bars to display and a
relatively small amount of space. This is less of a problem than it
sounds, as a negative value is still displayed (but presumably offset
than it should have been).

The PR also introduces a minimum width configuration option, set to 2
pixels by default.