Zoom for categoryAxisRenderer: manage the zoom - fix #187
Copy of orignal ticks and data is done in order to be able to restore them on zoom out.
Change of the ticks and the data to match the selected area.
Implementation of restoration on zoom out.

This code only works for x axis.
This code may be unefficient for large data (to be tested).